// src/store/main/type.ts
export interface IMainState {
  queryInfo: any
  usersList: any[]
  usersCount: number
  departmentList: any[]
  departmentCount: number
  menuList: any[]
  menuCount: number
  roleList: any[]
  roleCount: number
  goodsList: any[]
  goodsCount: number
  categoryList: any[]
  categoryCount: number
  storyList: any[]
  storyCount: number
  goodsCategoryCount: any[]
  goodsCategorySale: any[]
  goodsCategoryFavor: any[]
  goodsAddressSale: any[]
  goodsAmountList: any[]
}